requirements.txt
# Core dependencies for SIPD AI Chatbot
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI/ML dependencies
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Database and vector store
pgvector==0.2.4
psycopg2-binary==2.9.7
chromadb==0.4.18

# Data processing
json5==0.9.14
requests==2.31.0

# Modal.com deployment
modal==0.57.0

# HTTP client for async requests
aiohttp==3.9.1
aiofiles==23.2.0

# Additional utilities
python-dotenv==1.0.0
loguru==0.7.2
typing-extensions==4.8.0
jinja2==3.1.2
markdown==3.5.1